18 APR 2024 Create Folder Contracts which Holds Interfaces -->add IGenericRepository Interface to that folder which will declare the methods(No Implementation to be written) Create Folder Repositories/Services which holds Classes that will implement the Interfaces from Contracts -->add GenericRepository Class to Repositories/Services folder which will have the implementation/Definations for the Methods declared in IGenericRepository Interface  In this Class GenericRepository while we implement the methods, we should inherit this class from IGenericRepository  As we do operations in DB like we used DependencyInjection in LeaveTypeController class to use DB DependencyInjection we do the same here  Inject DependencyInjection using Constructor
We successfully Created Interface and Implemented the Interface in a Class>Now we have to Refactor our LeaveTypeController to use the Contractor to do the things
LeaveTypeController has the reponsiblity of doing CRUD operations for LeaveType module now by using Contractors that CRUD operations will take place
So we refactor our LeaveTypeController to use Contractor.

Create a ILeaveTypeRepository Interface in Contract which is dedicated to LeaveType module and this inherits IGenericRepository
Create a LeaveTypeRepository class which will implement the ILeaveTypeRepository Interface that Inherits ILeaveTypeRepository and GenericRepository<LeaveType>(This GenericRepository is inherited because we are using database to do operations)
Inject database Injection in LeaveTypeRepository class

-->We have created the Services/Contracts but to use this we have to register it in the Program.cs just like we registered Automapper and other stuff

We will remove ApplicationDbContext from LeaveTypeController class because the Controller will not do any CRUD operations so there is no use of it there.
So we inject the file which is going to do that job i.e ILeaveTypeRepository in place of ApplicationDbContext

Once we do the injection everything in Controller went into error so we change everything.

Refactoring:
_context to leaveTypeRepository